statystyki ładowanie, zapisywanie i aktualizowanie w setterach, dodatkowo bron ustawienie broni zeby statystyki sie aktualizowaly po zalozeniu

+++ Assets/Scripts/CellsVariant/Interfaces/ILivingThing.cs
@@ -4,11 +4,11 @@ public interface ILivingThing
 {
     int HealthPoints { get; set; }
     bool IsAlive { get; }
-    int Damage { get; }
+    float Damage { get; }
     GameObject Corpse_Sprite { get; }
     int MaxHealthPoints { get; }
 
-    void TakeDamage(int value, string source);
+    void TakeDamage(float value, string source);
     
     
 }
